hack(ui): test setup with Snowpack
All in all, I'm not sure that I'm going to follow up with this, it's interesting, but not quite ready for prime time yet. I don't want to be spending my time working on snowpack, I wan to just use it and move on.

That said, I's a really exciting project, and I'm keeping this around so that I can more easily try it out again later.